Move output staging to somewhere earlier to facilitate app-replacement staging
This involves moving appfuture creation earlier in the DFK submission code.
I think this is ok, although it means the app future exists with a task ID before
the task structure exists at all.

This might also form the base for some kind of join future that makes the overall
app fail if stageout fails (if that is desired?) - because that doesn't happen at
the moment. It's probably desirable in case anyone assuming that an app that has
finished means that you can access the output file URLs of that app (which I think
you can't assume so at the moment)

There is basically no testing of stage out behaviour in the test suite at the
moment, and it is awkward to set up manually as you need two globus endpoints.