make output datafutures complete when staging task is done, rather than when core task app code is done
prior to this commit, datafutures were created by bash/python app decorators, and configured to
complete at the point that the core task app completes.

this commit creates the datafutures as part of stageout setup, rather than as part of decorated
app invocation.

when stageout is inhibited, we still need to create the datafutures (as was happening in
the previous implementation where those futures were out of the stageout inhibition path)

when stageout is inhibited, a datafuture will be marked as complete as soon as its
app future is complete - because the stageout task completing is what means the
data future is complete