Investigate how apps cannot be chained using datafutures, because Files are resolved properly but datafutures aren't.
This patch so far puts in stricter type checking that only lets you specify a file or a str as a
dependency

Although sanitize_and_wrap replaces all input Futures with the contents of those futures before a
task starts to run, that happens much later (when all futures are resolved) and so that is not
happening with staging.

That's actually probably not something that should happen with staging anyway - staging-in tasks
should depend on the input data being ready, rather than waiting for all input futures to be
ready.

so this patch probably needs to fiddle round with those datafutures being input deps to stage-in

so the DataFuture needs to be an input=[df] to the stage in app, as well as that same stage in app
have the file contained in the df be the output of the app. in the non-df case, that should still
work with a file being both an input and an output: we're using the app to take a file, not do
parsl staging, and return the same file, except now we've messed around with the materialisation
of that - and that messing round with the materialisation is what staging is all about.

=====
other notes:
commit 42fbf6c23be5055fc17f76eef887fdb21b98f0dc
Author: Ben Clifford <benc@hawaga.org.uk>
Date: Tue Jul 9 18:00:59 2019 +0000

    hack on issue #778 - datafutures completing at wrong time

    they are marked as complete when an appfuture is marked as complete, which is when the underlying
    function has completed, ignoring whether any stageout has happened or not.

    potentially the whole appfuture should not be marked as complete until the stageouts have all
    happened: the two choices seem to be:

     1.  appfuture completes when everything to do with the app execution (including staging) is
         finished - for example, so we can wait on appfutures before shutting down the process

     2.  appfuture completes as soon as the result value is available, leaving other things that
         might happen (such as data staging) exposed through other futures (such as the existing
         but broken DataFutures)


    but at the very least, datafutures should not be marked as complete and returning their file
    object until such time as stageout has happened to the URL identified in the file object
    returned

    Before this patch, the bash and python apps create DataFutures with a parent of the newly
    submitted app, and then store those in the outputs array.

      i) why are they different? can I merge them into one/

     ii) can I make the DataFutures come into existence in the DFK, at the point that stageouts
         happen - with the *stageout* apps being the parent apps for the DataFutures

    iii) if so, does anyone care that the DataFuture parent is no longer the original app it was
         submitted with?

     iv) how does that interact with later in this patch stack where staging can happen potentially
         in multiple stages?